Move effects section to separate FXML and add additional Java audio devices
